Add OSPF support
Globably ready to be reviewed, it surfaced some miss-configuration from
prod. For example missing stub network one 1 of the 2 routers of a site,
or asymetric OSPF metrics on a link (240 vs 290).

Also adds by default a passive lo0.0 interface

We could then be more strict about it, for example if both sides are on
different sites then enable BFD, if one side is a MR then set the metric
to 20000.

Change-Id: Ib0f46e073ab7cd2acedc68664dcefc46dddaa3da