Initial templating for CR routing-options
Left to do: static routes.

Changes the structure of the lvs and lvs6 from lists of string to a dict
of sites. So we can only pick the local lvs prefixes to be used in
routing-options aggregate, but keep the whole list in firewall filters.

Changes the structure of bgp_out & bgp6_out from lists of strings to a
dict of dicts so we can specify the export policy to apply to the
prefix. This also avoids having to use something like "network_pop" :)

Change-Id: I65aa18893279822fad48d68bd82c1514e982ef33