wikireplicas: open the firewall for multiinstance databases
In order for the VM proxies to contact the rebuilt wikireplica cluster,
cloud networks must be open to ports 3311-3318. A firewall rule on ferm
will allow cloud public IPs to connect to them right now for testing,
but I have discovered that the network ACL appears to be blocking the
ports.

Bug: T271476
Change-Id: I51d7b142a4f5d13a181565a46a3788649fc14eb8