Add a mediawiki-api term to the analytics-in4 filter
The Analytics team runs an ETL process called Camus that fetches Kafka
topics and dumps the data on Hadoop HDFS. In order to do so for the
Eventgate streams, it needs to know their list from the Mediawiki API.
In the current setup the HTTP call is made via the webproxy, but we had
recently some trouble in importing data due to the webproxy down for an
outage. This change is made to bypass the webproxy and call directly the
api-ro discovery endpoint.

Bug: T274951
Change-Id: Ied2ce6a309e1b3d4ee821f013b4569c4076647ce