Enable per flow ECMP for kubernetes/kubestage
As a followup of 9546198, enable ECMP on kubernetes and kubestage
clusters across DCs. This is going to use a src+dst IP hashing scheme
as the rest of the configuration is already set for that. While this
isn't taking into consideration Layer 4 fields, it will probably do for
now as we don't know yet of a very high traffic single client for our
services.

As far as TCP session consistency goes, our services are stateless and
all have HTTP semantics, so the loss of a node with the resulting
rehashing and rerouting should lead to just an inevitable request
retry.

This has been tested on kubestage in codfw and seemed to work, but some
more extensive testing, especially simulating a node failure.

Bug: T238909
Change-Id: I920d4feb2be842abcfd0843f7a972ffbbbe8e9c4