Remove Anycast IPv4 ranges from bgp_out policy in eqord
The current policy for creating aggregate routes on CR routers
is done such that only routes originated locally at a given
site can contribute to generating the aggregates.

As eqord is only a network POP we will never have local routes
there which would trigger the aggregate to be created.  In that
case it doesn't make sense to have our anycast ranges listed in
the bgp_out policy applied to transit and peers.

Making this change so our config and intent is clearer.

Change-Id: Ifc1595dac6a3b57ab3ccbba3f1ecc0cff8d58ef3