Simplify labs-in4/6 firewall filters
By moving the relevant terms from external/dedicated filters to where
they're referenced.
That makes them easier to read as there is only 1 reference of each.
In preparation to moving them to Capirca

The now unused filters will need to be manually deleted from the
routers as the "replace:" is at their level.

Change-Id: I8966c8af55022b2d0749cc6d0ad767f53eb811e5