Adding 'quality-of-service' template for use on QFX/EX series switches.
Template is static, only configuring the shared-buffer partitioning,
which we want to allocate as much as possible to the 'lossy' class (as
all our traffic defaults to this).

I've set the egress multicast percent to 10, rather than the Juniper
generic reccomendation of 20.  I suspect that is more than enough for
what we might have there (OSPF?  VRRP?).

Added template in 'common' as I suspect we will want this on msw and
cloudsw also.  Only included in asw.conf for now.

Bug: T284592
Change-Id: I6862bc29372cb58f7e77c814b46a1cc51a225bfe