Added optional ability to enable uRPF filtering on arbitary CR ints
This should enable us to list interfaces under "urpf_strict" for a
given device to enable uRPF filtering (strict) for that interface.

Couldn't see any particularly good way to do it based on interface
description, so in the abscence of a way to mark them in netbox
took this approach.  Happy to discuss if this is the best way.

Rebased and will send its implementation in a following CR.

Bug: T285461
Bug: T298087
Change-Id: I0d38408a14813a9bac5bebd4287b24542db392d6