Add access to port 7000 on aqs_group temporarily
As part of the migration from Cassandra version 2 to 3, we need to use
sstableloader from a host in the analytics VLAN to TCP port 7000 on the
new AQS hosts.

The ferm rule has already been added, but this commit permits access
from the analytics VLAN to this port on the network devices as well.

It is only a temporary rule and will be removed when the migration has
been completed.

Bug: T291472
Change-Id: Ifae38a9928e8acbbe251c62d2c589b6493e07bad