Add policy-statement to CRs which sets next-hop self in iBGP.
This new policy will be applied with an export statement on the
local Confed_xxxxx BGP groups on CRs.  This will cause prefixes
announced to iBGP peers (i.e. the other CR) to have a next-hop
IP set to the interface IP being used for the BGP session itself.

This means CRs will not see the IP of the external BGP peer on
routes exchanged between them, but will rather see the IP of
the CR announcing the prefix itself.

The change means that CRs do not need to know about link addressing
on other CR devices.  But more importantly it should resolve some
edge-case niggles we have been having since adding the additional
port on at the Equinix IXP in Ashburn (eqiad).  These issues stem
from the fact both CRs are connected to the same peering LAN /
Ethernet segment.  This caused certain problems when iBGP routes
were learnt from the partner CR with next-hop on this subnet itself.

See T295650 and T295152 for detail of problems experienced.

Bug: T295672
Change-Id: I466453e623f1268646e142ac9088ea73966faaba