Modified loopback4 filter to allow NTP commands to run
Currently on our CR routers you cannot execute the show commands to
display the status of NTP associations on the router.  The reason
foir this is because the router makes a connection to itself to get
this info when the commands are executed, and this is being blocked
by our "loopback4" filter applied to lo0 interface.

See Juniper knowledge base article about this here:

https://kb.juniper.net/InfoCenter/index?page=content&id=KB11436

This change modifies the existing filter rule "allow_ntp4" so it accepts
packets with either source or destination port 123.  Previously it only
accepted packets to destination 123.

Bug: T296623
Change-Id: I03acde4a11c69a70bc6c1745549932adeef9a55f