Allow cloud-hosts1-eqiad DHCP responses to eqiad CRs
While investigating a problem with reimaging cloudvirt1028 it became
apparent that responses from install1003 to cr2-eqiad, for DHCP
requests that router had relayed on to it, were being blocked.

The reason for this is cr1-eqiad is VRRP master for the subnet
install1003 resides on, which meant the responses were routed back via
CR1.  Each CR, when relaying DHCP to the install server, sets the
source IP to its own IP on the interface the request was received on.
Responses from the install server duely come back to the CR IP of that
interface.

CR1 is VRRP master for the subnet install1003 is on, and thus all
responses it sends towards 10.64.20.x are sent via CR1.  cr1-eqiad,
receiving these packets destined for CR2's IP on the cloud-hosts1-eqiad
subnet, forwards them out it's own directly connected interface on that
subnet (xe-3/0/4.1118).  All well and good so far.

The problem is when the packets arrive in to cr2-eqiad on its
interface connected to that vlan (xe-3/0/4.1118).  The labs-in4 filter
has a term, deny-private-subnets, which blocks traffic coming in to that
interface with a destination inside 10.0.0.0/8.  This makes sense for
the expected traffic flow, we want to drop traffic coming from cloud
hosts with private destinations.

The problem here is the circular routing from cr1-eqiad to cr2-eqiad for
the DHCP responses.  This term, designed to block cloud hosts from
sending to 10.x addresses, blocks install1003 sending packets to
cr2-eqiad.

This change will allow DHCP response traffic from install1003 to route
into either CR from cloudsw on Vlan1118, which covers this edge case,
and means dhcp relay will work properly for this Vlan from both CRs.

Bug: T296906
Change-Id: Ic9cf717cba9448b8ec75def4575570c1f9085946