Updated iBGP policy to process local and remote routes differently.
This updates adds a new as-path definition, which matches prefixes
with the first AS in their path belonging to the range we use for
confederation sub-AS's.  In other words it matches BGP routes learnt
from other WMF sites over transport links.

This as-path is matched in a new policy-statement, IBGP-OUT.  The
intention is to discriminate between routes learnt from remote WMF
sites, and those learnt locally (either from peering/transit or from our
our own internal services like PyBal/anycast).

Routes from remote sites will be sent with their next-hop preserved
as what is learnt on the WAN.  i.e. the original next-hop from the
originating site.  This means the IGP cost will be effective in a
tie-break, as the cost to the next-hop reflects the OSPF cost to that
remote site.  Remote routes also have the BGP MED attribute set to
this IGP cost, which keeps existing setup introduced in T283163.

When applied this will mean remote WMF prefixes in iBGP are exchanged
with identical attributes to what they had been prior to the initial
change for next-hop self (see task).

It preserves setting next-hop to the routers own loopback for routes
*not* learnt from other WMF sites, however.  In other words the locally
learnt routes will still have next-hop set to the router's own loopback.
This retains the funcationality to support connectivity to the same IXP
LAN on both cr1-eqiad and cr2-eqiad, without hitting corner cases like
T295650.  It also ensures that if a local iBGP route is selected as best
by a router on a given site, traffic will always forward by the router
generating that update.

Bug: T295672
Change-Id: Ic83d6dea3336a951a8df5baf40d353d67b088e13