Removing entries from cr-analytics filter that refer to 'sodium'
The 'sodium' host was the old WMF debian repo mirror box.  It was
decomissioned earlier today as we are no longer using it.

Since then, however, Homer has been unable to generate the config
for cr1-eqiad.  This is because there is a Capirca defined filter
in use on the CR routers, analytics-in4, which refers to this host,
and Capirca cannot find it anymore as it's removed from Netbox.

As the host is gone the filter entries permitting traffic to it are
of no further use.  This change removes them so the filter will
still generate properly.

Change-Id: Ib26be8d50adc1514fd11cb4ea483f1f6fe723a18