Modify labs-in[4|6] filters in eqiad to allow traffic to codfw backups
The new cloudceph nodes and others on the cloud-hosts1-eqiad subnet
(vlan1118) in eqiad need to connect to backup hosts in codfw, but
currently this traffic is not permitted in either the v4 or v6 access
lists which are applied ingress on traffic from cloud hosts in eqiad.

This change adds a term to explicitly allow traffic from the cloud
subnet in eqiad to the two current backup hosts in codfw.

Change-Id: I007c0e9d091d39e7f526a22c0f3ffe55df4ab8b9