Add eBGP peering between CR routers and datacenter switches.
Submitting this as a proposed way to do some basic automation for the
CR to datacenter switch eBGP peerings.  I've still to submit changes
for the QFX devices, but I want to get this through as it involves
changes on live production routers.

The approach is not very sophisitcated, it definitely involves
duplication of some data that is in Netbox.  But it follows a similar
pattern that we have used elsewhere, and seems to me to be a
reasonable compromise between better automation and doing it manually.

Right now the import/export policies on the peering are set to NONE,
being cautious.  If this is approved/when BGP is up they can be
replaced with appropriate filters to allow the required routes to be
exchanged.

Mostly interested to get feedback on the approach or if we might
instead tackle this a different way.

Bug: T299758
Change-Id: I3d9f6cbb8cd4c9a09f6169d131058715a7bc8685