Add inbound and outbound BGP filters on CR to SPINE eBGP sessions
This change modifies the existing policy BGP_Default, adding a new
term to also match an IPv6 default route.  This policy is applied
outbound to the spine layer as we only wish to send a default route.

The CR also adds a new policy, BGP_Switch_In, which is intended to
control what prefixes are accepted from switches by the core routers.
In brief this is configured to:

1.  Allow prefixes matching the existing import policies for end-hosts
doing BGP.  i.e. we allow all the LVS, Anycast & K8s prefixes, using
the same policies that are already in plcae to correctly match IPs.

2.  Allow any prefixes from the 10.0.0.0/8 range which are length /24
or shorter (longer prefixes are not allowed, as we don't have smaller
subnets than /24, and we want to prevent leakage of EVPN MAC/IP
routes).

3.  Allow IPv6 prefixes from within the existing prefix-list
'private6', up to length /64 or shorter.  Longer prefixes than /64 are
also blocked similar to with v4.

Bug: T299758
Change-Id: Ib9aec7dacbed8940efdf57fe1ce9c7231d0b0188