Adjust CR templates so BGP_Switch_In doesn't reference K8s policy.
The new policy added earlier, BGP_Switch_In, referred to an import
policy which only exists in eqiad and codfw.

Specifically these policies relate to K8 hosts and the policy around
BGP for them.  The referenced policeis are added based on the whether
the 'customers' block is defined for a site, and if the specific K8s
ASNs are listed in it.  Right now only eqiad and codfw have this in
place.

For now taking the simple option and wrapping those terms in an 'if'
statement so they won't be added at other sites.

THIS CR JUST CORRECTS A TYPO IN PREVIOUS CHANGE 759503

Bug: 299758
Change-Id: I87a89215690b60bdc0db33d3eb8ae65a70b01929