Adjust CR Internal Anycast BGP Templates
This change adjusts the format for our Anycast peerings inside
sites.yaml, to allow for the specification of an IPv4 and IPv6 peer.
The template format is the same as we use for Kubernetes hosts to
specify a neighbor over each transport.

It will introduce a new BGP group - 'Anycast6' - but as the vars are
defined right now, with no IPv6 peers defined in 'anycast_neighbors'
for any site, the group will have no neighbors defined.  This will
change when v6 is enabled for Wikidough.

Mostly a no-op on the CRs.  It also adds a prefix-list to match the v6
range dedicated for Wikidough anycast IPs.  It adds a term to the
anycast_import policy which matches on this prefix list.  Same policy
can be used for the v4 and v6 peerings.

Tested on cr1-eqiad and works as expected, no diff in routes learnt
from existing Anycast peers with the policy adjusted.

I also removed the new /48 for Anycast services from the bgp6_out vars
for esams.  I think it's best we get the peering working internally,
verify the /128s are in the table and announced, before we announce
the range upstream / to peers.  #netops will raise a change for that
when the time comes.

Bug: T301165
Change-Id: I5380e33eaa739ab879993924920d032fb1be64ca