Modify vars for esams to announce IPv6 Anycast range
This change modifies the bgp6_out dict for esams to include the newly
assigned v6 prefix 2001:67c:930::/48.

We are learning /128 prefixes from doh VMs there via BGP, so with this
in place the aggregate /48 route should be created and thus announced
upstream to peers and transit.

I'm also correcting the 'anycast6' prefix list, the mask there needs
to match that on the public route.

Provided this works ok will engage with transits in AMS to accept the
route, and then submit follow up CR(s) for the other sites as they're
done.

Bug: T301165
Change-Id: I1ff66fb4bc887d885341315aef5c287daa99aab7