Change CR policy for creating aggregate Anycast routes
The introduciton of a routed access layer in drmrs and eqiad rows e
and f involves Anycast hosts BGP peering with their top-of-rack
switch, rather than directly peering with core routers.

The current aggregate route configuration for the Anycast /24 and /48
prefixes defines what routes need to be in the table for the aggregate
to be created (and thus announced publically), from a particular site.

The policy defined for this is currently matching on AS-path, and only
matching the Anycast ASN.  We wish to modify the policy to allow
Anycast routes generated locally to contribute to the aggregate, even
if the as-path contains one or more intermediate ASNs belonging to the
datacenter switches.

One way to accomplish this is by changing the policy as per this CR.
The approach is to first filter out all Anycast routes learnt from
remote WMF POPs (based on the AS number range we use for these), and
then allow all remaining Anycast routes (which should only be the
local ones).

The other option is obviously a per-site config at each POP to match
the particular AS-paths we expect to see at each, but that seems
more difficult to automate.

Bug: T302315
Change-Id: I29ed692ca099382aa5978ebecf2ad444d8e31b51