Initial changes to Homer config and templates for EVPN switches Eqiad
The changes here are sufficient to run homer against the new switches
installed in rows E/F eqiad resulting in a no-op.

It also allows DC-ops/other teams to run homer to configure the
switch ports as required based on the data in Netbox.

There is a good deal more config on these new switches, OSPF, BGP,
VXLAN etc., that is not covered in these limited changes, but will
not be overwritten by homer either.  I'm working through the config
for those based on feedback in
I83c308cfe2efa6a24113e7b711415f3425d8408c however I think it's best
to add these bits first, to unblock things while we get the strucure
right on the other templates.

Bug: T299758
Change-Id: I87fbfdf92d29b1f9839ae9f4e3ea9e5fd1723fa2