Re-enable direct path to Seabone / Telecom Italia in Eqiad
Our direct peering to TI was de-preffed a few weeks back as users
reported issues reaching Eqiad from them.

Today we have a different report from TI users in Milan.  We are
sending the traffic to Telia who then hand it off to TI in Ashburn.

This CR removes the de-pref of TI to place the traffic back on the
direct path between Wikimedia and TI, to see if that will improive
the situation.

Change-Id: Ie95ccb4c1c1e1ce3e9b576411abd4fe1c6e2b1c7