Modify cr-loopback Capirca definition to make it compatible with QFX
This change adjusts the cr-loopback Capirca policy so that what it
produces is compatible with the QFX5120 devices deployed as layer-3
switches in drmrs.  It produces a functionally equivalent, but very
slightly different filter than those currently in place on those
devices.  I'll detail the differences in more detail in the phab task.

It makes some smaller changes to the CRs, namely splitting the BGP into
two terms, so src-port and dst-port can be used, as 'port' on its own
isn't an option for TCP on the QFX switches.  Same with NTP.  It also
adds an additonal port to the DHCP terms to match what's already
configured in drmrs.

NOTE: Patchset3 changes 'port' to 'destination-port' for NTP and BFD
terms.  While the QFX accepts this syntax, it won't allow you to
apply the filter if any terms have just 'port' in them.  Should be
functionally the same for us.

Bug: T304553
Change-Id: I2e2b98fa9cda9d086f9b6f81ba94266d3cd9001c