Add template to configure IPv6 RAs on CRs and L3 Switches
This automates the config of the IPv6 RAs on required interfaces on
both CR routers and L3 switches.

For L3 switches the logic is as simple as adding it for all IRB
interfaces, which works fine with our current usage of IRBs there.

On CRs this causes a few changes, but should be fine:

1. It adds RA config for ae1.102 on the esams CRs.  This is the
sandbox Vlan and doesn't strictly need it, but won't cause an issue
to add.

2. It adds the 'virtual-router-only' command to the config in eqsin,
which brings it in line with our other sites.  It was probably just
an oversight that it wasn't there before.

3. Many of the RA config stanzas have the interface IP configured
after the 'prefix' statement.  This jinja template instead puts the
network address there.  As seen before both of these configs produce
the same result.  We can be confident that it will work on MX's as we
currently have a mixture of network_addr and interface_addr defined on
them for different Vlans.

Bug: T299758
Change-Id: I5ab553fe492fc006b483c6d82bad912dfbfe6183