Add inbound filter to analytics IRB interfaces on EVPN switches Eqiad
As discovered by ayounsi, adding uRPF filtering on an interface
blocks DHCP requests, as they are IPv4 packets with source 0.0.0.0.

That problem was overcome for MX routers in change
I0ca7feec98ef498c5ba0ac7815919a2167d83f17, however it turns out the
'exception' command is not available on QFX switches (see T285461).

This change removes the uRPF config from Analytics vlans on EVPN
based devices, and adds a filter for those vlans on the new switches
installed in Eqiad row E/F instead.  The filter allows DHCP requests
specifically for v4, as well as allowing traffic from 'private'
source IPs usind the existing capirca definitions.

While this is not as targetted as the uRPF filter, in terms of the
goal to prevent data exfiltration to the internet from these hosts it
should serve the same purpose.  Traffic from any private IP will be
dropped on the CRs.

Bug: T299758
Change-Id: I917be852dd745b4381f500a9f562aae3875534e0