Modify homer automation for IPv6 RAs to allow for custom interfaces
The logic to indentify interfaces that need to send IPv6 RAs in the
previous template is not sufficient to cover all our use cases.

Specifically we have it enabled on fxp0.0 (mgmt) interface on the CRs
in drmrs.  To allow for this I've added a new optional field for
devices.yaml, 'ip6_ra', under which you can list interfaces to
simply list under 'router-advertisement', without any prefix set up
matching what is on the drmrs routers.

Bug: T299758
Change-Id: Ie46c796d66f002816d1bb0c443adb4721b7a1656