Modify CR loopback filter and add VRF-specific filter for switches
This change modifies the cr-loopback filter to include the required
changes so the same one can be used on EVPN switches.  Namely a
prefix-list (bgp-sessions-overlay) is defined with an apply-path
statement.  That won't match anything on CRs but will on the EVPN
devices.

Secondly loopback4/6 prefix list is added to the NTP term as the
switches have a private IP on their loopback which otherwise would
be blocked when running 'show ntp' commands.

Lastly a new filter is defined to be applied to loopback interface
'units' bound to a non-default routing-instance/vrf.  Initially I
had tried to use the same filter on these but the QFX5120s have
insufficient TCAM to support, so a shorter list is needed there.

Bug: T304553
Change-Id: Ie67a1c46fb1394c07f819816e69f68d4d6bcdd40