Move elements from CR BGP policy and group config to separate files
This change moves policy elements, prefix-lists and policy-statements
specifically, which are common to both l3 switches and core routers
to separate files.  This allows them to be 'included' in the
templates for those separate platforms, and avoids us repeating the
same or almost identical code in two sets of templates.

The same goes for some of the BGP groups that are defined, in
particular those which connect to end servers, i.e. PyBal, Anycast,
Kubernetes.  I've created a new folder for these snippets.

I'll follow-up with the automation for the Leaf switches which will
include the same files, but want to make sure the approach is valid
for CRs first.  Right now it's a no-op on the CRs apart from the
order of the prefix-lists changing.

Change-Id: Id44614e4a12fc97a2d9abfa9f4bfc9cdb2eac387