Add template for custom routing-instances on cloudsw
Patch adds new vars to be added to define VRFs and their assciated
IDs, as well as place specific interfaces on devices into those VRFs
(routing instances).

Does not use 'replace' for the whole routing-instance stanza, to
facilitate protocols to be configured in it manually for the time
being.

Also includes changes / vars for the new cloudsw in rows E/F and
the existing ones to move the new xlink irb sub-ints to vrf 'cloud',
which will be required for the migration and also gives an example
of how it would be used.

Diff on current cloudsw: https://phabricator.wikimedia.org/P27958

Bug: T304989
Change-Id: I47f3657d57d0e7b38cfb3b6efb3b11c0abeb500c