Adjust custom-vrf template to support devices with 32-bit ASNs
Using a 32-bit ASN in VPN address family route-target communities is
supported in Junos, but the AS number needs to be suffixed with 'L'
if one is used.

Patch modifies the custom-vrf template to add this L if required.

Also adding the per-device ASNs for new cloudsw in rows E/F Eqiad,
as well as for drmrs switches.  Patch is a no-op on these devices
right now when run.

Bug: T304989
Change-Id: If9a3ebfe408f4fd80e2e8b974955091f57e1b551