Adjust switch_neighbors vars to include peer-as of other side
This patch adds an optional variable to the "dual_af" family vars
which can be used to include the neighbours ASN if required.

For instance the 'switch_neighbors' group peers with both the EVPN
swithes and cloud switches in Eqiad, which use different ASNs.

Also including the updated definition for the LSW neighbours and
the new cloudsw.

Bug: T304936
Change-Id: I4831588530a314396d1fd553abd55fac02d9ee9c