Remove DHCP option 82 insertion for all l3_switches
With the introduction of the cloud vrf to separate cloud and
production realms on the cloud switches, the various 'cloud-hosts'
vlans now have their default gateway on the cloudsw themselves.

As such the top-of-rack switches themselves need to perform the full
DHCP relay operation, sending requests to the install sever when
end hosts make DHCP requests.  The configuration to enable this
also includes the DHCP Option 82 insertion, which means that the
Vlan config to only insert the Option 82 string, without forwarding
the request, is not needed.

This patch changes the vlan template so cloudsw devices are no longer
processed differently than other devices of type l3_switch.

Bug: T304936
Change-Id: Ie0657cfcaa571b96dc36b7bf80cd2cc930a3965c