Move cloudsw capirca def to roles and asn to devices
All cloudsw are now layer-3 switches and need the loopback filter
applied.  So makes more sense to move this to the 'role' section
for all cloudsw.

The eBGP config means they do not all share the same ASN now, so
making the opposite change there to specify this on a per-device
basis.

Bug: T304936
Change-Id: Id3f5f82565ed6746e119f755cf0bf070645f2055