Add static routes to those exported by cloudsw to CR routers
Givent the cloudsw2 devices have no BGP license there are static
routes in place for their loopback on the upstream cloudsw1
switches in the same rack.  As there is a single LAG between these
devices there is no need for a dynamic protocol in terms of failover.

The statics need to be announced to the CR routes to make them more
widely reachable.

Bug: T304936
Change-Id: Ib411348ed7ed60ff1b1dfa48bde7a23585ccf4a9