Modify Eqiad CR labs-in filter to allow BGP and ICMP
The labs-in filter needs to be applied ingress from production realm
networks on cloudsw devices in eqiad.  The filter/ruleset is largely
unchanged from what was previously configured on xe-3/0/4.1118 on
each.

The differences here are to allow BGP (cloudsw now peer with the CRs)
and remove the VRRP section (VRRP is only internal on the cloudsw so
doesn't need to be allowed out to CRs).

Additionally I've allowed ICMP for v4 and v6, as during the migration
it complicated things not being able to send pings to test
connectivity.

Bug: T304989
Change-Id: I6a2242303f92ce286c0580605465b0f44327ac05