Adjust EVPN switch BGP template
Small adjustment that wasn't caught as part of
Iea2e6c859380288d05d50eea60b4feb0dae4ba93.  The first 'if' statement
here would return true if k8s_mlstaging_neighbors is defined at a
higher level of the hierarchy (for instance defined for site), and
the 'set' command would then fail if no per-device k8s_mlstaging
neighbors were defined.

The 'if' should only be triggered if per-device neighbors are
defined, the 'set' is used to enable re-use of the same template
as on the CRs.

Bug: T302198
Change-Id: Ie3737c8e9e99d30b20d98408aa569192cd14a0ea