Include routes from proto aggregate in Cloud VRF out filter to CRs
Adding 'proto aggregate' to the source protocols included in the
Cloud_outfilter policy used on cloudsw1-c8/d5.  This controls what
routes are sent to the CRs, and following the removal of static
route for WMCS /24 range, and use of aggregate, this is needed
to ensure the full /24 is announced to the edge.

No-op on switches (change already done/tested manually).

Bug: T315956
Change-Id: I24d648124988e5490d011016f355e8770e7b5fd2