Change CR template to only include Kubedse in BGP_Switch_in for eqiad
The change that Ben added yesterday was well structured, however the
BGP_Switch_In policy ended up being slightly incorrect.  All other
K8s groups exist in Codfw and Eqiad, and there is an 'if' surrounding
the includes which only adds them for those sites.

The new Kubedse seems only to exist in Eqiad, and as a result the
policy from 64609.policy template is not added in Codfw.  This
was causing an issue when the policy name was referenced in the
BGP_Switch_In policy.

Patch changes the conditionals so Kubedse is only included in the
BGP_Switch_In policy in Eqiad.

Bug: T310174
Change-Id: I6a56c2204ca88fc23e30e8ccd82642908ac11ccf