Add explicit BFD session mode (single/multi-hop) to Anycast groups
We had an incident previously where BFD sessions were failing in
drmrs as the switches were trying to do multi-hop BFD over UDP
port 4784, but the VMs anycasting were using single-hop over UDP
3784.

Today we had the opposite, whereby the core routers in ulsfo were
trying to do single-hop BFD, and the VMs were using multi-hop.

In both cases Bird on the VM/server side appeared to be using the
correct mode based on whether it's BFD peer was on its local
subnet or not.

I am not 100% sure of what logic JunOS uses to select this, an
obvious pattern is not clear.  But either way this change to our
templates should explicitly set the mode on the Juniper side in
both cases, so we have no further incidents like this.

Tested on cr4-ulsfo and it fixed issue, existing BFD sessions were
not affected by the config addition.

Bug: T304501
Change-Id: I0ba8dab4a8f2a4c7e5a7b5942828d631131ae559