Management routers: replace bootp with dhcp-relay
And enable it on all POPs

The bootp feature has been deprecated with the recent upgrades and is
not working in its curent form. As dhcp-relay is the future deploy it
now.

We should also evaluate doing the same for the core routers. bootp is
not deprecated there but dhcp-relay exists so that would help
consistency.

Bug: T271583
Change-Id: I9008aac76ec19de5b0e8435949a142e8ba28153e