Change hard-coded loopback int references in EVPN templates
I'd changed the loopback interface name for the production vrf
on the lsw devices to match the way it was done on cloudsw.

This updates the device config and routing-instance template to
match.  Ultimately won't need to have this hard coded, this change
is just to remove the current diff.

Bug: T312635
Change-Id: I5113bcdd6497ab8b6d445dc49ed749fccc7f3ded