Move if statement around 'ospf' section in asw template
I'd moved the if statement within the 'ospf' section on this change
previously, the suggestion being this would remove the section for
devices that didn't match the condition of being 'evpn'.

Now that it's merged I see it's too early to do this, as the OSPF
config used on the l3switches in drmrs would be removed.

This change moves the 'if' back outside the 'ospf' header so that
the section won't be cleared for the drmrs switches for now.

A little further out we need to decide whether or not to use OSPF
for the directly connected networks in drmrs or not.  There may be
an argument for only using BGP on those switches and not two
protocols.

Bug: T312635
Change-Id: I8c57ba2d9a7fb801bb87115d614c0a8d5876495c