Remove VRF-specific loopback filter from row E/F switches
The loopback filter on lo0.5000 on the row E/F (EVPN) switches caused
a bit of a problem as it was seemingly preventing the switches from
generating ICMPv6 TTL Exceeded messages correctly.

Based on Juniper tech-note advice (see task) this patch will remove
this filter entirely, so the device falls back on using the filter
defined on lo0.0 for all routing-instances.

I will do a further review to make sure that the lo0.0 filter covers
everything we need to for the VRF.  As all IPs on these boxes are
not routable from the internet there is not much danger for now
anyway.

Bug: T324033
Change-Id: Iaec413aa69fdeef8d20f42604ef687b5e21dd43d