Remove atlas-ulsfo from cr-border-in.pol as it's not live
Capirca filter generation for CR border-in is failing after the Netbox
Capirca definitions were updated today.  This seems to be due to a
reference to atlas-ulsfo which is now not found, as that device was
set to decomissioned status late last year.

Removing the reference to that from the policy for now to resolve the
failure.

Change-Id: Idad807fadf22ec56c0b48a3249d9ae0c4aeeaf44