Disable Telemetry on eqsin switches
eqsin switches are running a code version with the bug documented in
PR1566528, causing mostly logs flood. We curently don't need that
feature and will need to upgrade the switches the day we do.

There are multiple ways of implementing the fix:
 1) This one, fastest and most specific
 2) Having it behind a feature flag: cleaner but more complex
 3) Deploying it to all the switches altogether

I'm fine using 2 or 3 if anyone have a strong preference

Bug: T316532
Change-Id: I844a46428ff0a910c51968eec11fddaf470c2ac2