Adjust EVPN BGP type-5 route creation / export to include host routes
This change means that EVPN-enabled switches will export type-5
host routes from their local EVPN DB if the host routes are known
on an 'Anycast GW' interface.  In other words if the host routes
belong to a stretched Vlan that spans multiple switches.

The local EVPN DB is populated from ARP/ND tables as bindings are
added/removed.

Without this just the Vlan prefix itself (i.e. 10.192.0.0/22) is
exported by participating LEAFs.  Which means the SPINE does not
see which particular switch a given IP is reachable via.  Traffic
works in that scenario, but hops via a random LEAF first, before
getting re-transmitted to the correct one.  This change ensures
the optimal path is visible to the SPINE.

Diff from previous patch set is we now filter on interface, so we
continue to only export the vlan prefix (i.e. 10.64.1.0/24) for
networks that do not stretch over multiple switches.

Bug: T329369
Change-Id: I1086f0d4978cc26a4b5bfa53d7e78aa7bd918c3e