Add automation for EVPN BGP peerings
Adds a new section to 'sites' to define the list of participating
EVPN devices at a site, divided into two groups, the devices acting
as route reflectors and their clients.

If we ever have more than 1 IGP / RR cluster we can further group
by the cluster_ids, and assign devices to a specific cluster, but
for now we've only 1 at each so that'd be overkill I think.

No-op on current EVPN switches.

I also removed the schema definition for "vrf_ints" as we are now
using the Netbox field for that rather than the YAML.

Bug: T327934
Change-Id: Ie635d6c51ec6bb9d007637f9f43e88750da4ed4c