Homer changes as part of WMCS codfw migration to cloudsw1-b1-eqiad
These changes reflect the current state of play on the core routers
in codfw, which have had a few manual changes in past hours as part
of a managed migration for WMCS from asw-b1-codfw to
cloudsw1-b1-codfw.

Some further changes will be required, I will submit more patches as
I go, although some need to be done manually because the timing of
changes has to be carefully managed.  I'll do my best to keep on
top to avoid causing problems for others reconfigring the CRs
using homer.

Bug: T327919
Change-Id: I9e8aacac0d11de7b84f052c5ed7ae0f22f5a1048