Move cloudsw prefix-list filters from templates to YAML
As per discussion on Id39ed0c6e8340870eef75b5329d80885e2d3c520. this
attempts to move the definition of ranges cloudsw will announce at
a given site from templates to our YAML files.

For discussion, there are obviously different ways to approach this,
we may wish to nest the definition, or define it at a different
level of hierarchy.

Bug: T327919
Change-Id: Ia5fd1a9a4b2499f5c0d049bc5f53133b98df3bfe