Add EVPN protocol config for enabled L3 switches
Adding base evpn protocol definition and VNI<->Vlan bindings which
should have been included with previous patch to configure EVPN BGP
protocol itself.  Also including the forwarding-options stanza
to ensure ECMP works in both overlay and underlay routing tables.

Following comment relating to drmrs I had a look there and have
adjusted the config to 'replace' the routing-options for all
L3 devices.  This does involve a few changes, but they are ok to
make.

In 'drmrs' it adds the ecmp-fast-reroute command, which is a good
thing to have (although with small number of routes convergence
is fast so not too important).  It also removes some aggregate
and static route config there that is not required on the L3
switches.

Diff: https://phabricator.wikimedia.org/P46105

Almost a no-op on the EVPN switches in eqiad, it will remove stale
static route for old netflow server.

Bug: T327934
Change-Id: I0e784ea5671f6bd6353fa6541c32ebe139c8c3da