Add generic way to create static routes on switches
Adds a generic structure that can be used to define static routes
that are required on L3 switches.  We don't have many of these,
specifically the cloudsw2 devices require some in eqiad.  Which
seems enough of a special-case not to be worth doing anything more
than a generic method for statics.

There is also an issue with encoding the IPv6 default in the YAML,
which is discussed in full in the task.  But for now this works
fine and will clear the homer diff mails.

Bug: T334281
Change-Id: I891beff85db3dc91a943904ec18d3dec6dc2c1b8