Automate and update DHCP relay configuration
This config adds automation for the 'forwarding-options' config
section for L3 switches, which previously had this done manually.

It adds both the hashing config (see T230955), and the DHCP
relay configuration.

The changes also adjust how the CR DHCP relay function is configured,
moving from the legacy 'bootp helper' config to use the same
'dhcp relay' syntax as deployed on the switches.

Other than the ECMP change on L3 switches the patch makes no
functional change to behaviour of any device, however the naming of
certain elements have been changed.

I changed the strucuture of the 'dhcp_server' var in YAML so the name
of the install server could be used as the server-group name.  I
think this improves the readability of the resulting config.

Depends on Iab89a60a90855ee3fcc8d716a644a3fd8e19aed5 to expose VRF
membership of interfaces so they can be excluded on cloudsw, selected
on evpn switches.

Diff that this will produce can be seen here:

https://phabricator.wikimedia.org/P46663

The jinja is a little verbose in the dhcp_server.conf file.  I did
play with making a function in the Netbox wmf plugin instead, but the
code there ended up messier for several reasons so I went back to the
Jinja.

Bug: T312635
Bug: T320508
Bug: T328737
Change-Id: If0a55fa9b9ea038e77ceb896c7a0046f24cae884